+++ spec/spec_helper.rb
@@ -35,8 +35,8 @@ require 'water_drop'
 # Configure for test setup
 WaterDrop.setup do |config|
   config.send_messages = false
-  config.connection_pool_size = 1
-  config.connection_pool_timeout = 1
-  config.kafka.hosts = ['localhost:9092']
+  config.connection_pool.size = 1
+  config.connection_pool.timeout = 1
+  config.kafka.seed_brokers = ['localhost:9092']
   config.raise_on_failure = true
 end